refactor(concentrator): split routes.ts into route modules
routes.ts: 1964 -> 331 lines (-83%). Extracted into src/concentrator/routes/:
- sessions.ts (270) - /sessions/* endpoints
- admin.ts (368) - users, shares, inter-session links
- api.ts (527) - push, crashes, files, settings, transcribe, misc
- stats.ts (141) - cost/analytics
- spawn.ts (91) - /api/spawn, /api/dirs
- blob-store.ts (268) - blob + shared-files store
- shared.ts (127) - route helpers factory

Pure structural move, no behavior changes. Permission checks preserved
in-place per handler. routes.ts remains composition root.
